+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 + [x] Update `accordion` for themes
 + [x] Update `calendar` for themes
 + [x] Update `button-group` for themes
++ [x] Update `checkbox` for themes
 + [x] Added class `error`
 + [x] Updated `breadcrumbs default`, added new style `breadcrumbs arrow`